Added two Supertest + Jest API tests: get stone by ID and create new stone. All passed.

+++ backend/tests/setup.js
@@ -10,7 +10,9 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
+  if (mongoose.connection.readyState !== 0) {
+    await mongoose.connection.dropDatabase();
+    await mongoose.connection.close();
+  }
   await mongoServer.stop();
-});
+}, 15000);
